fix(autoupdate): Fix version number comparison.
The output directory format changed recently to include the channel as a
prefix. This broke devserver's logic to parse the version out of the
directory name which manifested itself by stumbling over the + character
that now became part of the version string. Instead read the entire
version string from version.txt and fix the version comparison code to
not care about + or any other non-integer character in the version.

Also, one shouldn't write C++ code in Python, that just wastes everyones
time. ;-)