emerge-gitclone: new tool for cloning overlays into dev images
Portage 2.2 supports syncing portage trees via git but it will only do a
git pull, never a git clone to initialize the repos for the first time.

This little tool fills that gap, the name is related to a similar tool
`emerge-webrsync` which fetches the Gentoo portage tree as a tarball
snapshot over HTTP. It is only necessary to run this once in a new
developer image or container, after that use `emerge --sync`.