Add support for the new image format
Images with the kernel as a separate payload need to have a different app ID
in order to ensure that we only pass them to versions of Update Engine that
will understand the additional payload. Add support to devserver to detect
whether the client is asking for a new image or a legacy one, and add
support to cros_generate_update_payload to build an appropriate image.