emerge-gitclone: Attempt to check out the exact release commit
Previously, the Git clone would check out the branch, which could
have advanced quite a bit beyond the currently running release.  It
will now try to find the specific commit in the manifest used to
build the release.  If that fails, it will fall back to using the
release branch again, which is still usually close enough.